requirements.txt
# Web frameworks
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.40

# ML & NLP
scikit-learn==1.3.2
mlflow==2.11.1
transformers==4.37.2
torch==2.0.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
urduhack==1.1.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Data & plotting
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0
